version bump

+++ bin/stop.sh
@@ -17,6 +17,8 @@
 #
 # Author/Maintainer: Konrad Bächler <konrad@diva.exchange>
 #
+# -e  Exit immediately if a simple command exits with a non-zero status
+set -e
 
 PROJECT_PATH="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"/../
 cd ${PROJECT_PATH}
@@ -24,12 +26,18 @@ PROJECT_PATH=`pwd`/
 
 echo -n "Terminating "`pgrep -f "^node .*${PROJECT_PATH}dist/main.js$" | wc -l`" processes..."
 
-pkill -SIGTERM -f "^node .*${PROJECT_PATH}dist/main.js$"
+if [[ `pgrep -f "^node .*${PROJECT_PATH}dist/main.js$"` ]]
+then
+  pkill -SIGTERM -f "^node .*${PROJECT_PATH}dist/main.js$"
+fi
 
 while [[ `pgrep -f "^node .*${PROJECT_PATH}dist/main.js$"` ]]
 do
-  echo -n `pgrep -f "^node .*${PROJECT_PATH}dist/main.js$" | wc -l`"."
-  sleep 2
+  if [[ `pgrep -f "^node .*${PROJECT_PATH}dist/main.js$"` ]]
+  then
+    echo -n "."
+    sleep 2
+  fi
 done
 
-echo "...done"
+echo "done"
